delete filters file
